Add code to address HASS startup problem
HASS currently appears to ignore retained messages on the topics it subscribes to when it starts up.

So, though Door 1 and Door 2 retained status messages may be present on the status topics, HASS ignores these and the door state in HASS may be incorrect (in the case of the cover platform, the state will be "unknown").

This is an attempt to workaround the problem. HASS can publish a birth message when it starts up. The code added in this commit subscribes to the birth_topic and listens for the payload on that topic. When that message is received, GarHAge will publish current status for Door 1 and Door 2 (if enabled).