Add door status check before executing OPEN or CLOSE action (#6)
* Add string variables and code to update them

Added string variables for each door to hold a string representation of the current state of the door to be set by the publish_doorX_status() functions, and the code to each function to set the strings at the same time that an update is published over MQTT.

* Added sanity-checking function

Added sanity checking function for each garage door to check whether the requested state on the action topic matches the current state of the door; added these functions to the logic in triggerDoorAction() so that an action command only triggers a door action if the door is NOT already in the requested state.

* Changed logic processing to allow for status updates

Modified the way the doorX_sanityCheck() boolean is tested to allow for publishing a status update when the sanityCheck fails (i.e. the requested state and current state are equal) as this could be a sign that either the broker or client are out of sync with GarHAge.